chore: move trait implementation
