+++ core/src/models/exif_meta/mod.rs
@@ -1,11 +1,8 @@
 pub mod db;
+pub mod str;
 
-use once_cell::sync::Lazy;
-use regex::Regex;
 use serde::{Deserialize, Serialize};
-use snafu::prelude::*;
 use sqlx::FromRow;
-use std::str::FromStr;
 use strum_macros::Display as EnumDisplay;
 
 #[derive(Clone, Debug, Serialize, FromRow)]
@@ -31,49 +28,4 @@ pub enum Maker {
     Konica,
     #[strum(serialize = "CANON")]
     Canon,
-}
-
-impl FromStr for Maker {
-    type Err = Error;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        static RE: Lazy<Regex> = Lazy::new(|| Regex::new(r"([a-zA-Z]+)").unwrap());
-        let caps = RE.captures(s).context(NotValidSnafu {
-            name: s.to_string(),
-        })?;
-
-        match caps[0].to_lowercase().as_str() {
-            "fujifilm" => Ok(Maker::Fujifilm),
-            "konica" => Ok(Maker::Konica),
-            "canon" => Ok(Maker::Canon),
-            _ => Err(Error::NotValid {
-                name: s.to_string(),
-            }),
-        }
-    }
-}
-
-#[derive(Debug, Snafu, PartialEq)]
-pub enum Error {
-    #[snafu(display("Invalid Maker: {}", name))]
-    NotValid { name: String },
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_parses_fujifilm() {
-        assert_eq!(Maker::from_str("Fujifilm"), Ok(Maker::Fujifilm));
-    }
-
-    #[test]
-    fn it_parses_konica() {
-        assert_eq!(Maker::from_str("Konica"), Ok(Maker::Konica));
-    }
-
-    #[test]
-    fn it_parses_canon() {
-        assert_eq!(Maker::from_str("Canon"), Ok(Maker::Canon));
-    }
 }
