+++ core/src/models/exif_meta/str/mod.rs
@@ -1,0 +1,1 @@
+pub mod maker;
